requirements.txt
beautifulsoup4
googlesearch-python
openai
pathlib
PyPDF2
requests
selenium
tensorflow
tokenizers
transformers